backend/pyproject.toml
[project]
name = "backend"
version = "0.0.1"
description = ""
authors = [{ name = "Hai Binh Do", email = "binhdoitsme@gmail.com" }]
dependencies = [
    "uvicorn>=0.23.2",
    "fastapi>=0.103.1",
    "websockets>=11.0.3",
    "aiokafka>=0.8.1",
    "pydantic-core>=2.10.0",
    "sqlalchemy>=2.0.21",
    "psycopg2-binary>=2.9.9",
]
requires-python = ">=3.11"
readme = "README.md"
license = { text = "MIT" }

[tool.pdm.scripts]
start = { cmd = "uvicorn main:app --host=0.0.0.0", env_file = ".env" }

[tool.pdm.dev-dependencies]
dev = ["types-confluent-kafka>=1.0.0"]
